requirements.txt
discord-py-interactions
Flask
